Added support for the Tags API call.
Added support for the Tags API call and did a slight refactoring of the
statistics. Tags/Info returns a stats property, so I changed
MandrillStats to be the aggregate stats instead of MandrillUserStats
which was just abstracting the aggregate stats. Also, moved the stats
and aggregate stats up a level in the structure.